+++ include/vrv/view.h
@@ -10,6 +10,7 @@
 
 #include "devicecontextbase.h"
 #include "scoredef.h"
+#include "smufl.h"
 #include "vrvdef.h"
 
 namespace vrv {
@@ -20,6 +21,7 @@ class BarLine;
 class Beam;
 class BeamDrawingParams;
 class Breath;
+class ControlElement;
 class Chord;
 class ControlElement;
 class DeviceContext;
@@ -40,6 +42,8 @@ class Lb;
 class Measure;
 class MNum;
 class Mordent;
+class Nc;
+class Neume;
 class Num;
 class Octave;
 class Options;
@@ -55,6 +59,7 @@ class Slur;
 class Staff;
 class Svg;
 class Syl;
+class Syllable;
 class System;
 class SystemElement;
 class Tempo;
@@ -172,7 +177,8 @@ protected:
     ///@{
     void DrawSystem(DeviceContext *dc, System *system);
     void DrawSystemList(DeviceContext *dc, System *system, const ClassId classId);
-    void DrawScoreDef(DeviceContext *dc, ScoreDef *scoreDef, Measure *measure, int x, BarLine *barLine = NULL);
+    void DrawScoreDef(DeviceContext *dc, ScoreDef *scoreDef, Measure *measure, int x, BarLine *barLine = NULL,
+        bool isLastMeasure = false);
     void DrawStaffGrp(DeviceContext *dc, Measure *measure, StaffGrp *staffGrp, int x, bool topStaffGrp = false,
         bool abbreviations = false);
     void DrawStaffDef(DeviceContext *dc, Staff *staff, Measure *measure);
@@ -180,8 +186,8 @@ protected:
     void DrawStaffDefLabels(DeviceContext *dc, Measure *measure, ScoreDef *scoreDef, bool abbreviations = false);
     void DrawBracket(DeviceContext *dc, int x, int y1, int y2, int staffSize);
     void DrawBrace(DeviceContext *dc, int x, int y1, int y2, int staffSize);
-    void DrawBarLines(DeviceContext *dc, Measure *measure, StaffGrp *staffGrp, BarLine *barLine);
-    void DrawBarLine(DeviceContext *dc, int y_top, int y_bottom, BarLine *barLine);
+    void DrawBarLines(DeviceContext *dc, Measure *measure, StaffGrp *staffGrp, BarLine *barLine, bool isLastMeasure);
+    void DrawBarLine(DeviceContext *dc, int y_top, int y_bottom, BarLine *barLine, bool eraseIntersections = false);
     void DrawBarLineDots(DeviceContext *dc, StaffDef *staffDef, Staff *staff, BarLine *barLine);
     void DrawLedgerLines(DeviceContext *dc, Staff *staff, ArrayOfLedgerLines *lines, bool below, bool cueSize);
     void DrawMeasure(DeviceContext *dc, Measure *measure, System *system);
@@ -352,6 +358,16 @@ protected:
     ///@}
 
     /**
+     * @name Methods for drawing neumes.
+     * Defined in view_neume.cpp
+     */
+    ///@{
+    void DrawSyllable(DeviceContext *dc, LayerElement *element, Layer *layer, Staff *staff, Measure *measure);
+    void DrawNc(DeviceContext *dc, LayerElement *element, Layer *layer, Staff *staff, Measure *measure,
+        wchar_t fontNo = SMUFL_E990_chantPunctum, int xOffset = 0, int yOffset = 0);
+    void DrawNeume(DeviceContext *dc, LayerElement *element, Layer *layer, Staff *staff, Measure *measure);
+    ///@}
+    /**
      * @name Methods for drawing Floating child classes.
      * They are base drawing methods that are called directly from DrawFloatingElement.
      * Call appropriate method of child classes (Slur, Tempo, Tie, etc).
@@ -413,13 +429,8 @@ protected:
      * Defined in view_mensural.cpp
      */
     ///@{
-    void DrawMensuralStem(DeviceContext *dc, Note *note, Staff *staff, data_STEMDIRECTION dir, int radius,
-        int xn, int originY, int heightY = 0);
-    void DrawMensurCircle(DeviceContext *dc, int x, int yy, Staff *staff);
-    void DrawMensurDot(DeviceContext *dc, int x, int yy, Staff *staff);
-    void DrawMensurHalfCircle(DeviceContext *dc, int x, int yy, Staff *staff);
-    void DrawMensurReversedHalfCircle(DeviceContext *dc, int x, int yy, Staff *staff);
-    void DrawMensurSlash(DeviceContext *dc, int x, int yy, Staff *staff);
+    void DrawMensuralStem(DeviceContext *dc, Note *note, Staff *staff, data_STEMDIRECTION dir, int radius, int xn,
+        int originY, int heightY = 0);
     void DrawMaximaToBrevis(DeviceContext *dc, int y, LayerElement *element, Layer *layer, Staff *staff);
     void CalculateLigaturePosX(LayerElement *element, Layer *layer, Staff *staff);
     void DrawProportFigures(DeviceContext *dc, int x, int y, int num, int numBase, Staff *staff);
@@ -438,8 +449,10 @@ protected:
      * Defined in view_graph.cpp
      */
     ///@{
-    void DrawVerticalLine(DeviceContext *dc, int y1, int y2, int x1, int nbr);
-    void DrawHorizontalLine(DeviceContext *dc, int x1, int x2, int y1, int nbr);
+    void DrawVerticalLine(DeviceContext *dc, int y1, int y2, int x1, int width, int dashLength = 0);
+    void DrawHorizontalLine(DeviceContext *dc, int x1, int x2, int y1, int width, int dashLength = 0);
+    void DrawVerticalSegmentedLine(DeviceContext *dc, int x1, SegmentedLine &line, int width, int dashLength = 0);
+    void DrawHorizontalSegmentedLine(DeviceContext *dc, int y1, SegmentedLine &line, int width, int dashLength = 0);
     void DrawSmuflCode(
         DeviceContext *dc, int x, int y, wchar_t code, int staffSize, bool dimin, bool setBBGlyph = false);
     void DrawThickBezierCurve(DeviceContext *dc, Point bezier[4], int thickness, int staffSize, float angle = 0.0);