+++ src/view_control.cpp
@@ -424,7 +424,7 @@ void View::DrawOctave(
 
     /************** parent layers **************/
 
-    LayerElement *start  = dynamic_cast<LayerElement *>(octave->GetStart());
+    LayerElement *start = dynamic_cast<LayerElement *>(octave->GetStart());
     LayerElement *end = dynamic_cast<LayerElement *>(octave->GetEnd());
 
     if (!start || !end) {
@@ -432,11 +432,16 @@ void View::DrawOctave(
         return;
     }
 
-    /********** adjust the end position ***********/
+    /********** adjust the start / end positions ***********/
+
+    if ((spanningType == SPANNING_END) || (spanningType == SPANNING_MIDDLE)) {
+        x1 += (m_doc->GetGlyphWidth(SMUFL_E0A2_noteheadWhole, staff->m_drawingStaffSize, false) / 2);
+    }
 
     if ((spanningType == SPANNING_START_END) || (spanningType == SPANNING_END)) {
-        if (octave->HasEndid())
+        if (octave->HasEndid()) {
             x2 += (m_doc->GetGlyphWidth(SMUFL_E0A2_noteheadWhole, staff->m_drawingStaffSize, false) / 2);
+        }
     }
 
     /************** draw it **************/
@@ -534,6 +539,7 @@ void View::DrawSlur(DeviceContext *dc, Slur *slur, int x1, int x2, Staff *staff,
     data_STEMDIRECTION startStemDir = STEMDIRECTION_NONE;
     data_STEMDIRECTION endStemDir = STEMDIRECTION_NONE;
     data_STEMDIRECTION stemDir = STEMDIRECTION_NONE;
+    bool isGraceToNoteSlur = false;
     int y1 = staff->GetDrawingY();
     int y2 = staff->GetDrawingY();
 
@@ -575,20 +581,21 @@ void View::DrawSlur(DeviceContext *dc, Slur *slur, int x1, int x2, Staff *staff,
         endStemDir = endChord->GetDrawingStemDir();
     }
 
-    Layer *layer1 = NULL;
-    Layer *layer2 = NULL;
+    if (startNote && endNote && startNote->IsGraceNote() && !endNote->IsGraceNote()) {
+        isGraceToNoteSlur = true;
+    }
 
     // For now, with timestamps, get the first layer. We should eventually look at the @layerident (not implemented)
     // if (start->Is(TIMESTAMP_ATTR))
     //    layer1 = dynamic_cast<Layer *>(staff->FindChildByType(LAYER));
     // else
-    layer1 = dynamic_cast<Layer *>(start->GetFirstParent(LAYER));
+    Layer *layer1 = dynamic_cast<Layer *>(start->GetFirstParent(LAYER));
 
     // idem
     // if (end->Is(TIMESTAMP_ATTR))
     //    layer2 = dynamic_cast<Layer *>(staff->FindChildByType(LAYER));
     // else
-    layer2 = dynamic_cast<Layer *>(end->GetFirstParent(LAYER));
+    Layer *layer2 = dynamic_cast<Layer *>(end->GetFirstParent(LAYER));
 
     assert(layer1 && layer2);
 
@@ -644,6 +651,10 @@ void View::DrawSlur(DeviceContext *dc, Slur *slur, int x1, int x2, Staff *staff,
     if (slur->HasCurvedir()) {
         drawingCurveDir
             = (slur->GetCurvedir() == curvature_CURVEDIR_above) ? curvature_CURVEDIR_above : curvature_CURVEDIR_below;
+    }
+    // grace notes - always below unless we have a drawing stem direction on the layer
+    else if (isGraceToNoteSlur && (layer1->GetDrawingStemDir(startNote) == STEMDIRECTION_NONE)) {
+        drawingCurveDir = curvature_CURVEDIR_below;
     }
     // the normal case
     else if (slur->HasDrawingCurvedir()) {
@@ -767,8 +778,19 @@ void View::DrawSlur(DeviceContext *dc, Slur *slur, int x1, int x2, Staff *staff,
             }
         }
         else {
+            if (isGraceToNoteSlur) {
+                if (endNote) {
+                    y2 = endNote->GetDrawingY();
+                    x2 -= m_doc->GetDrawingUnit(staff->m_drawingStaffSize) * 2;
+                    isShortSlur = true;
+                }
+                else {
+                    y2 = y1;
+                }
+            }
             // (_)d
-            if (endStemDir == STEMDIRECTION_up) y2 = end->GetDrawingBottom(m_doc, staff->m_drawingStaffSize);
+            else if (endStemDir == STEMDIRECTION_up)
+                y2 = end->GetDrawingBottom(m_doc, staff->m_drawingStaffSize);
             // P(_)P
             else if (isShortSlur) {
                 y2 = end->GetDrawingBottom(m_doc, staff->m_drawingStaffSize);
@@ -857,7 +879,7 @@ void View::DrawSlur(DeviceContext *dc, Slur *slur, int x1, int x2, Staff *staff,
 
     // the normal case or start
     if ((spanningType == SPANNING_START_END) || (spanningType == SPANNING_START)) {
-        start->FindAllChildByAttComparison(&artics, &matchType);
+        start->FindAllChildByComparison(&artics, &matchType);
         // Then the @n of each first staffDef
         for (articIter = artics.begin(); articIter != artics.end(); ++articIter) {
             Artic *artic = dynamic_cast<Artic *>(*articIter);
@@ -877,7 +899,7 @@ void View::DrawSlur(DeviceContext *dc, Slur *slur, int x1, int x2, Staff *staff,
     }
     // normal case or end
     if ((spanningType == SPANNING_START_END) || (SPANNING_END)) {
-        end->FindAllChildByAttComparison(&artics, &matchType);
+        end->FindAllChildByComparison(&artics, &matchType);
         // Then the @n of each first staffDef
         for (articIter = artics.begin(); articIter != artics.end(); ++articIter) {
             Artic *artic = dynamic_cast<Artic *>(*articIter);
@@ -962,7 +984,7 @@ float View::AdjustSlur(Slur *slur, Staff *staff, int layerN, curvature_CURVEDIR 
     FindTimeSpanningLayerElementsParams findTimeSpanningLayerElementsParams;
     findTimeSpanningLayerElementsParams.m_minPos = p1->x;
     findTimeSpanningLayerElementsParams.m_maxPos = p2->x;
-    std::vector<AttComparison *> filters;
+    ArrayOfComparisons filters;
     // Create ad comparison object for each type / @n
     // For now we only look at one layer (assumed layer1 == layer2)
     AttNIntegerComparison matchStaff(STAFF, staff->GetN());
@@ -1229,7 +1251,7 @@ void View::AdjustSlurPosition(Slur *slur, ArrayOfLayerElementPointPairs *spannin
 
     int dist = abs(p2->x - p1->x);
     float posXRatio = 1.0;
-    
+
     ArrayOfLayerElementPointPairs::iterator itPoint;
     for (itPoint = spanningPoints->begin(); itPoint != spanningPoints->end();) {
         int y = BoundingBox::CalcBezierAtPosition(bezier, itPoint->second.x);
@@ -1620,9 +1642,9 @@ void View::DrawSylConnectorLines(DeviceContext *dc, int x1, int x2, int y, Syl *
             margin = (dist - ((nbDashes - 1) * dashSpace)) / 2;
         }
 
-        int i, x;
+        int i;
         for (i = 0; i < nbDashes; ++i) {
-            x = x1 + margin + (i * dashSpace);
+            int x = x1 + margin + (i * dashSpace);
             x = std::max(x, x1);
 
             DrawFilledRectangle(dc, x - halfDashLength, y, x + halfDashLength, y + width);
@@ -2138,7 +2160,7 @@ void View::DrawPedal(DeviceContext *dc, Pedal *pedal, Measure *measure, System *
 
     // just as without a dir attribute
     if (!pedal->HasDir()) return;
-    
+
     dc->StartGraphic(pedal, "", pedal->GetUuid());
 
     int x = pedal->GetStart()->GetDrawingX() + pedal->GetStart()->GetDrawingRadius(m_doc);
@@ -2158,11 +2180,12 @@ void View::DrawPedal(DeviceContext *dc, Pedal *pedal, Measure *measure, System *
         x -= bounceOffset.m_width;
     }
     if (pedal->GetDir() != pedalLog_DIR_up) {
-        if (pedal->GetFunc() == "sostenuto") code = SMUFL_E659_keyboardPedalSost;
-        else code = SMUFL_E650_keyboardPedalPed;
+        if (pedal->GetFunc() == "sostenuto")
+            code = SMUFL_E659_keyboardPedalSost;
+        else
+            code = SMUFL_E650_keyboardPedalPed;
     }
     str.push_back(code);
-    
 
     std::vector<Staff *>::iterator staffIter;
     std::vector<Staff *> staffList = pedal->GetTstampStaves(measure);
@@ -2206,11 +2229,11 @@ void View::DrawTempo(DeviceContext *dc, Tempo *tempo, Measure *measure, System *
     // First try to see if we have a meter sig attribute for this measure
     MeasureAlignerTypeComparison alignmentComparison(ALIGNMENT_SCOREDEF_METERSIG);
     Alignment *pos
-        = dynamic_cast<Alignment *>(measure->m_measureAligner.FindChildByAttComparison(&alignmentComparison, 1));
+        = dynamic_cast<Alignment *>(measure->m_measureAligner.FindChildByComparison(&alignmentComparison, 1));
     if (!pos) {
         // if not, try to get the first beat element
         alignmentComparison.SetType(ALIGNMENT_DEFAULT);
-        pos = dynamic_cast<Alignment *>(measure->m_measureAligner.FindChildByAttComparison(&alignmentComparison, 1));
+        pos = dynamic_cast<Alignment *>(measure->m_measureAligner.FindChildByComparison(&alignmentComparison, 1));
     }
     // if we found one, use it
     if (pos) {